SmearSlide tweaks: allow floating point percentages, require non-empty percentage fields to submit form, change "Name" to "Slide Name", "Id" to "ID"

+++ grails-app/domain/edu/umn/laccore/tmi/SmearSlideComponent.groovy
@@ -7,7 +7,7 @@ package edu.umn.laccore.tmi
 // rename to SedimentComponent?
 class SmearSlideComponent {
 	SedimentComponent component
-	int percentage
+	float percentage
 	boolean deleted
 	boolean withComponent = false
